Multiple improvements: Beryl's core had to be rewritten.

+++ src/coremods/core_dbmanager/cmd_dbadmin.cpp
@@ -19,6 +19,7 @@
 #include "managers/tests.h"
 #include "converter.h"
 #include "engine.h"
+#include "managers/settings.h"
 #include "core_dbmanager.h"
 
 CommandPWD::CommandPWD(Module* Creator) : Command(Creator, "PWD", 0)
@@ -28,7 +29,7 @@ CommandPWD::CommandPWD(Module* Creator) : Command(Creator, "PWD", 0)
 
 COMMAND_RESULT CommandPWD::Handle(User* user, const Params& parameters)
 {  
-       if (user->current_db)
+       if (user->GetDatabase())
        {
             const std::string path = user->current_db->GetPath().c_str();
             user->SendProtocol(BRLD_CURRENT_DIR, path, Daemon::Format("%s", path.c_str()).c_str());
@@ -39,6 +40,25 @@ COMMAND_RESULT CommandPWD::Handle(User* user, const Params& parameters)
        return FAILED;
 }
 
+CommandDefault::CommandDefault(Module* Creator) : Command(Creator, "DEFAULT", 0)
+{
+
+}
+
+COMMAND_RESULT CommandDefault::Handle(User* user, const Params& parameters)
+{  
+       std::string dbname = STHelper::Get("dbconf", "dbdefault");
+       
+       if (dbname.empty())
+       {
+             user->SendProtocol(ERR_NO_DEFAULT, dbname);
+             return FAILED;
+       }
+       
+       user->SendProtocol(BRLD_DB_DEFAULT, dbname);
+       return SUCCESS;
+}
+
 CommandDB::CommandDB(Module* Creator) : Command(Creator, "DB", 0)
 {
 
@@ -46,9 +66,9 @@ CommandDB::CommandDB(Module* Creator) : Command(Creator, "DB", 0)
 
 COMMAND_RESULT CommandDB::Handle(User* user, const Params& parameters)
 {  
-       if (user->current_db)
+       if (user->GetDatabase())
        {
-             std::string dbname = user->current_db->GetName();                                  
+             std::string dbname = user->GetDatabase()->GetName();                                  
              user->SendProtocol(BRLD_DB_NAME, dbname);
              return SUCCESS;	
        }
@@ -64,22 +84,24 @@ CommandChange::CommandChange(Module* Creator) : Command(Creator, "CHANGE", 1, 1)
 
 COMMAND_RESULT CommandChange::Handle(User* user, const Params& parameters)
 {  
-       const std::string& dbname = parameters[0];
-       std::shared_ptr<UserDatabase> database = Kernel->Store->DBM->Find(name);
+       std::string dbname = parameters[0];
+       std::shared_ptr<UserDatabase> database = Kernel->Store->DBM->Find(dbname);
 
        if (!database)
        {
-             user->SendProtocol(ERR_DB_NOT_FOUND, dbname);
+             user->SendProtocol(ERR_DB_NOT_FOUND, DB_NULL);
              return FAILED;
        }
-       
+
+       user->SetDatabase(database);
+       STHelper::Set("dbuser", user->login, dbname);
        user->SendProtocol(BRLD_DB_CHANGED, PROCESS_OK);
        return SUCCESS;
 }
 
 CommandDBLIST::CommandDBLIST(Module* Creator) : Command(Creator, "DBLIST", 0, 0)
 {
-      requires = 'm';
+      requires = 'r';
 }
 
 COMMAND_RESULT CommandDBLIST::Handle(User* user, const Params& parameters)
@@ -91,11 +113,11 @@ COMMAND_RESULT CommandDBLIST::Handle(User* user, const Params& parameters)
       for (DataMap::const_iterator it = dbases.begin(); it != dbases.end(); ++it)
       {
             std::shared_ptr<UserDatabase> udb = it->second;
-            
             const std::string& dbname = udb->GetName();
             const std::string& dbpath = udb->GetPath();
             Dispatcher::CondList(user, BRLD_DB_ITEM, dbname, dbpath);
       }     
+      
 
       Dispatcher::JustAPI(user, BRLD_DB_END);
       
@@ -104,20 +126,42 @@ COMMAND_RESULT CommandDBLIST::Handle(User* user, const Params& parameters)
 
 CommandDBCreate::CommandDBCreate(Module* Creator) : Command(Creator, "DBCREATE", 1, 1)
 {
-      requires = 'm';
+      requires = 'r';
       syntax = "<name>";
 }
 
 COMMAND_RESULT CommandDBCreate::Handle(User* user, const Params& parameters)
 {
-      const std::string& dbname = parameters[0];
+      std::string dbname = parameters[0];
+
+      /* 'dbdefault' is a reserved database name. */
+      
+      if (dbname == "dbdefault" || dbname == "core")
+      {
+             user->SendProtocol(ERR_DB_EXISTS, PROCESS_FALSE);
+             return FAILED;
+      }
+      
+      std::transform(dbname.begin(), dbname.end(), dbname.begin(), ::tolower);
       
       if (Kernel->Store->DBM->Create(dbname, dbname))
       {			
-          Kernel->Store->DBM->Load(dbname, dbname);
-
-          user->SendProtocol(BRLD_DB_CREATED, PROCESS_OK);
-          return SUCCESS;
+             Kernel->Store->DBM->Load(dbname);
+             user->SendProtocol(BRLD_DB_CREATED, PROCESS_OK);
+             sfalert(user, NOTIFY_DEFAULT, "Added database: %s", dbname.c_str());      
+             
+             if (user->current_db == nullptr)
+             {
+                  std::shared_ptr<UserDatabase> database = Kernel->Store->DBM->Find(dbname);
+                  
+                  if (database)
+                  {
+                        user->SetDatabase(database);
+                        STHelper::Set("dbuser", user->login, dbname);
+                  }
+             }
+             
+             return SUCCESS;
       }
       
       user->SendProtocol(ERR_DB_EXISTS, PROCESS_FALSE);
@@ -126,28 +170,45 @@ COMMAND_RESULT CommandDBCreate::Handle(User* user, const Params& parameters)
 
 CommandDBDelete::CommandDBDelete(Module* Creator) : Command(Creator, "DBDELETE", 1, 1)
 {
-      requires = 'm';
+      requires = 'r';
       syntax = "<name>";
 }
 
 COMMAND_RESULT CommandDBDelete::Handle(User* user, const Params& parameters)
 {
-      const std::string& dbname = parameters[0];
+      std::string dbname = parameters[0];
       
-       std::shared_ptr<UserDatabase> database = Kernel->Store->DBM->Find(dbname);
+      std::shared_ptr<UserDatabase> database = Kernel->Store->DBM->Find(dbname);
 
-       if (!database)
-       {
-             user->SendProtocol(ERR_DB_NOT_FOUND, dbname);
+      if (!database)
+      {
+             user->SendProtocol(ERR_DB_NOT_FOUND, PROCESS_NULL);
              return FAILED;
-       }
+      }
       
+      /* We cannot remove a database whenever is locked. */
+      
+      if (database->IsClosing())
+      {
+             user->SendProtocol(ERR_DB_BUSY, DATABASE_BUSY);
+             return FAILED;
+      }
+      
+      database->SetClosing(true);
+
       const std::string& dbpath = database->GetPath();
       
       if (Kernel->Store->DBM->Delete(dbname))
       {                 
-            user->SendProtocol(BRLD_DB_DELETED, PROCESS_OK);
+            if (database == user->GetDatabase())
+            {
+                  user->SetNullDB();
+            }
+            
             FileSystem::remove_directory(dbpath.c_str());
+            user->SendProtocol(BRLD_DB_DELETED, PROCESS_OK);
+            sfalert(user, NOTIFY_DEFAULT, "Removed database: %s", dbname.c_str());
+            
             return SUCCESS;
       }
       
@@ -164,5 +225,30 @@ COMMAND_RESULT CommandDBTest::Handle(User* user, const Params& parameters)
 {
       user->SendProtocol(BRLD_TEST, PROCESS_OK);
       TestHelper::Dump(user);
+      return SUCCESS;
+}
+
+CommandDBSetDefault::CommandDBSetDefault(Module* Creator) : Command(Creator, "SETDEFAULT", 1, 1)
+{
+      requires = 'r';
+      syntax = "<db name>";
+}
+
+COMMAND_RESULT CommandDBSetDefault::Handle(User* user, const Params& parameters)
+{
+      const std::string& dbname = parameters[0];
+
+      std::shared_ptr<UserDatabase> database = Kernel->Store->DBM->Find(dbname);
+
+      if (!database)
+      {
+             user->SendProtocol(ERR_DB_NOT_FOUND, PROCESS_NULL);
+             return FAILED;
+      }
+
+      sfalert(user, NOTIFY_DEFAULT, "Setting default database to: %s", database->GetName().c_str());
+
+      DBManager::SetDefault(dbname);
+      user->SendProtocol(BRLD_DEFAULT_CHANGED, PROCESS_OK);
       return SUCCESS;
 }
